requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
discord.py
frozenlist
idna
multidict
mysql-connector-python
propcache
py-cord
pytest
python-dotenv
yarl